+++ src/background/widgets/weg/instance.rs
@@ -3,7 +3,6 @@ use seelen_core::state::{FancyToolbarSide, HideMode, SeelenWegSide};
 use tauri::{WebviewWindow, Wry};
 use windows::Win32::{
     Foundation::{HWND, RECT},
-    Graphics::Gdi::HMONITOR,
     UI::WindowsAndMessaging::SWP_ASYNCWINDOWPOS,
 };
 
@@ -13,7 +12,7 @@ use crate::{
     log_error,
     state::application::FULL_STATE,
     widgets::{toolbar::FancyToolbar, WebviewArgs},
-    windows_api::{AppBarData, WindowsApi},
+    windows_api::{monitor::Monitor, AppBarData, WindowsApi},
 };
 
 pub struct SeelenWeg {
@@ -83,27 +82,27 @@ impl SeelenWeg {
         Ok(weg)
     }
 
-    pub fn get_weg_size_on_monitor(monitor: HMONITOR) -> Result<i32> {
+    pub fn get_weg_size_on_monitor(monitor: &Monitor) -> Result<i32> {
         let state = FULL_STATE.load();
-        let settings = &state.settings.by_widget.weg;
-        let monitor_dpi = WindowsApi::get_monitor_scale_factor(monitor)?;
-        let text_scale_factor = WindowsApi::get_text_scale_factor()?;
-        let total_size = (settings.total_size() as f64 * monitor_dpi * text_scale_factor) as i32;
+        let settings: &seelen_core::state::SeelenWegSettings = &state.settings.by_widget.weg;
+        let total_size = (settings.total_size() as f64 * monitor.scale_factor()?) as i32;
         Ok(total_size)
     }
 
-    pub fn set_position(&mut self, monitor: HMONITOR) -> Result<()> {
+    pub fn set_position(&mut self, monitor: &Monitor) -> Result<()> {
         let state = FULL_STATE.load();
         let toolbar_config = &state.settings.by_widget.fancy_toolbar;
+        let is_toolbar_enabled = state.is_bar_enabled_on_monitor(&monitor.stable_id2()?);
+
         let settings = &state.settings.by_widget.weg;
 
         let hwnd = HWND(self.hwnd()?.0);
-        let monitor_info = WindowsApi::monitor_info(monitor)?;
+        let monitor_info = WindowsApi::monitor_info(monitor.handle())?;
 
         self.theoretical_rect = monitor_info.monitorInfo.rcMonitor;
         self.webview_rect = monitor_info.monitorInfo.rcMonitor;
 
-        if toolbar_config.enabled && toolbar_config.hide_mode != HideMode::Always {
+        if is_toolbar_enabled && toolbar_config.hide_mode != HideMode::Always {
             let toolbar_size = FancyToolbar::get_toolbar_height_on_monitor(monitor)?;
             match state.settings.by_widget.fancy_toolbar.position {
                 FancyToolbarSide::Top => {
@@ -157,7 +156,7 @@ impl SeelenWeg {
         if self.webview_rect == WindowsApi::get_outer_window_rect(hwnd)? {
             return Ok(()); // position is ok no need to reposition
         }
-        self.set_position(WindowsApi::monitor_from_window(hwnd))?;
+        self.set_position(&Monitor::from(WindowsApi::monitor_from_window(hwnd)))?;
         Ok(())
     }
 }